+++ share/objects/unused-objects.lua
@@ -1,0 +1,6 @@
+local map = ...
+local objects = map.objects
+
+-- Used in old range check.
+objects ['n00D'] = nil
+objects ['n006'] = nil